Handle start/end positions equal to the root's text content length
Re-implement the `toRange` function to better handle cases where the
start or end positions include the end of the root element's text, or
the start or end positions are outside the range `[0, root.textContent.length]`.

This fixes an issue encountered in Hypothesis' PDF annotation where
creating an annotation would fail if the selection included the last
character of a PDF page's text layer.

This implementation also makes it more obvious which of the two possible
range configurations is used when the `start` or `end` offset occurs at
a text node boundary. In that case, the range could use either the end
of one text node or the start of the next as the boundary point. The
boundary point is always placed at the end of the earlier node in that case.